Several changes related to SDR1000 tx mode. Now sending foward and reflected power from the SDR1000 back to QtRadio via spectrum data. Added TX power and swr meter to QtRadio. Cleaned up serveral compiler warning for DttSP and dspserver.

+++ trunk/src/DttSP/banal.c
@@ -172,7 +172,7 @@ char *
 fmt_tv (struct timeval *tv)
 {
 	static char buff[256];
-	snprintf (buff, sizeof (buff), "%ds%du", tv->tv_sec, tv->tv_usec);
+    snprintf (buff, sizeof (buff), "%ds%du", (int)tv->tv_sec, (int)tv->tv_usec);
 	return buff;
 }
